Move MultiSpan to DataFormats/Common
