+++ RecoHGCal/TICL/plugins/TracksterLinksProducer.cc
@@ -1,7 +1,7 @@
 // Author: Felice Pantaleo, Wahid Redjeb (CERN) - felice.pantaleo@cern.ch, wahid.redjeb@cern.ch
 // Date: 12/2023
 #include <memory>  // unique_ptr
-#include "DataFormats/HGCalReco/interface/MultiVectorManager.h"
+#include "DataFormats/Common/interface/MultiSpan.h"
 #include "FWCore/Framework/interface/stream/EDProducer.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
@@ -203,11 +203,11 @@ void TracksterLinksProducer::produce(edm::Event &evt, const edm::EventSetup &es)
   auto resultMask = std::make_unique<std::vector<float>>(original_global_mask);
 
   std::vector<edm::Handle<std::vector<Trackster>>> tracksters_h(tracksters_tokens_.size());
-  MultiVectorManager<Trackster> trackstersManager;
+  edm::MultiSpan<Trackster> trackstersManager;
   for (unsigned int i = 0; i < tracksters_tokens_.size(); ++i) {
     evt.getByToken(tracksters_tokens_[i], tracksters_h[i]);
-    //Fill MultiVectorManager
-    trackstersManager.addVector(*tracksters_h[i]);
+    //Fill MultiSpan
+    trackstersManager.add(*tracksters_h[i]);
   }
 
   // Linking
