+++ Validation/HGCalValidation/interface/BarrelVHistoProducerAlgo.h
@@ -23,7 +23,7 @@
 #include "DataFormats/ParticleFlowReco/interface/PFRecHit.h"
 
 #include "RecoLocalCalo/HGCalRecAlgos/interface/RecHitTools.h"
-#include "DataFormats/HGCalReco/interface/MultiVectorManager.h"
+#include "DataFormats/Common/interface/MultiSpan.h"
 
 #include "SimDataFormats/CaloAnalysis/interface/CaloParticle.h"
 #include "SimDataFormats/Vertex/interface/SimVertex.h"
@@ -238,7 +238,7 @@ public:
                                       unsigned int layers,
                                       const ticl::RecoToSimCollection& recSimColl,
                                       const ticl::SimToRecoCollection& simRecColl,
-                                      MultiVectorManager<reco::PFRecHit> const& barrelHits) const;
+                                      edm::MultiSpan<reco::PFRecHit> const& barrelHits) const;
   void layerClusters_to_SimClusters(const Histograms& histograms,
                                     const int count,
                                     edm::Handle<reco::CaloClusterCollection> clusterHandle,
@@ -251,7 +251,7 @@ public:
                                     unsigned int layers,
                                     const ticl::RecoToSimCollectionWithSimClusters& recSimColl,
                                     const ticl::SimToRecoCollectionWithSimClusters& simRecColl,
-                                    MultiVectorManager<reco::PFRecHit> const& barrelHits) const;
+                                    edm::MultiSpan<reco::PFRecHit> const& barrelHits) const;
 
   void tracksters_to_SimTracksters_fp(const Histograms& histograms,
                                       const int count,
@@ -270,7 +270,7 @@ public:
                                 std::vector<SimVertex> const& simVertices,
                                 unsigned int layers,
                                 std::unordered_map<DetId, const unsigned int> const&,
-                                MultiVectorManager<reco::PFRecHit> const& barrelHits) const;
+                                edm::MultiSpan<reco::PFRecHit> const& barrelHits) const;
   void fill_generic_cluster_histos(const Histograms& histograms,
                                    const int count,
                                    edm::Handle<reco::CaloClusterCollection> clusterHandle,
@@ -283,7 +283,7 @@ public:
                                    unsigned int layers,
                                    const ticl::RecoToSimCollection& recSimColl,
                                    const ticl::SimToRecoCollection& simRecColl,
-                                   MultiVectorManager<reco::PFRecHit> const& barrelHits) const;
+                                   edm::MultiSpan<reco::PFRecHit> const& barrelHits) const;
   void fill_simCluster_histos(const Histograms& histograms,
                               std::vector<SimCluster> const& simClusters,
                               unsigned int layers) const;
@@ -299,7 +299,7 @@ public:
                                          unsigned int layers,
                                          const ticl::RecoToSimCollectionWithSimClusters& recSimColl,
                                          const ticl::SimToRecoCollectionWithSimClusters& simRecColl,
-                                         MultiVectorManager<reco::PFRecHit> const& barrelHits) const;
+                                         edm::MultiSpan<reco::PFRecHit> const& barrelHits) const;
   void fill_cluster_histos(const Histograms& histograms, const int count, const reco::CaloCluster& cluster) const;
 
   double distance2(const double x1, const double y1, const double x2, const double y2) const;
@@ -309,7 +309,7 @@ public:
 
   DetId findmaxhit(const reco::CaloCluster& cluster,
                    std::unordered_map<DetId, const unsigned int> const&,
-                   MultiVectorManager<reco::PFRecHit> const& hits) const;
+                   edm::MultiSpan<reco::PFRecHit> const& hits) const;
 
   struct detIdInfoInCluster {
     bool operator==(const detIdInfoInCluster& o) const { return clusterId == o.clusterId; };
