+++ Validation/HGCalValidation/interface/BarrelValidator.h
@@ -32,7 +32,7 @@
 #include "SimDataFormats/Associations/interface/LayerClusterToSimClusterAssociator.h"
 #include "SimDataFormats/Associations/interface/TICLAssociationMap.h"
 
-#include "DataFormats/HGCalReco/interface/MultiVectorManager.h"
+#include "DataFormats/Common/interface/MultiSpan.h"
 
 class PileupSummaryInfo;
 
@@ -66,7 +66,7 @@ public:
                                 std::vector<size_t>& selected_cPeff,
                                 unsigned int layers,
                                 std::unordered_map<DetId, const unsigned int> const&,
-                                MultiVectorManager<reco::PFRecHit> const& barrelHits) const;
+                                edm::MultiSpan<reco::PFRecHit> const& barrelHits) const;
 
   static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
